refactor: Move and rename GeoCard in TripMap
